Keylogger/requirements.txt
pynput
pyperclip
pillow
scapy
win32gui
pygetwindow